refactor(postgrest): remove the key-path filter surface
Deletes `PostgrestTypedQuery+Filters.swift`. All three readers of the key-path
spelling have moved: `where`/`order`, `@SelectionOf`'s `selectString`, and
`PostgrestUpdate`'s subscript. There is now exactly one way to write a filter:
`.where { $0.column.eq(value) }` in place of `.eq(\.column, value)`, and
`.order { $0.column.desc() }` in place of `.order(\.column, ascending:)`.

`@Table` stops generating `columnName(for:)` and `PostgrestRelation` drops the
requirement, which also removes the generated `default: fatalError` — a key
path to a computed property was a runtime trap and is now a compile error.
Hand-written conformances declare a `Columns` namespace instead.

`PostgrestKeyPathFilterable` is renamed `PostgrestFilterableRequest`, since it
no longer has anything to do with key paths, and moves next to `where(_:)`.
`limit(_:)` moves alongside `order(_:)`.

Not marked breaking: the typed surface is new in the unreleased v3 and has
never shipped, so there is no released call site to migrate.
