+++ Sources/PostgREST/Query/PostgrestTypedQuery+Where.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension PostgrestKeyPathFilterable {
+extension PostgrestFilterableRequest {
   /// Scopes the request by a filter.
   ///
   /// The closure receives the relation's column namespace. Each PostgREST operator is a method
@@ -98,5 +98,12 @@ extension PostgrestTypedQuery where Phase: PostgrestTransformablePhase {
     return PostgrestTypedQuery<R, Output, PostgrestTransformPhase>(
       builder: PostgrestRequestBuilder(carryingFrom: builder, request: request)
     )
+  }
+
+  /// Limits the number of rows returned.
+  ///
+  /// Like either `order(_:)` overload this moves the request into ``PostgrestTransformPhase``.
+  public func limit(_ count: Int) -> PostgrestTypedQuery<R, Output, PostgrestTransformPhase> {
+    PostgrestTypedQuery<R, Output, PostgrestTransformPhase>(builder: builder.limit(count))
   }
 }
