+++ Tests/PostgRESTTests/PostgrestFilterTests.swift
@@ -24,13 +24,6 @@ struct PostgrestFilterTests {
     }
 
     static let columns = Columns()
-
-    static func columnName(for keyPath: PartialKeyPath<Self>) -> String {
-      switch keyPath {
-      case \Self.id: "id"
-      default: fatalError("unmapped key path")
-      }
-    }
   }
 
   private func rendered(_ filter: PostgrestFilter<Todo>) -> [String] {
